backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==22.1
djoser==2.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0